move components to their folder, follow conventions
move components to their folder, follow conventions
