+++ client/components/Footer.js
@@ -1,4 +1,4 @@
-import styles from "../../styles/Footer.module.css";
+import styles from "../styles/Footer.module.css";
 
 export default function Footer() {
     return (
