+++ client/components/HeadComponent.js
