+++ client/components/Navbar.js
@@ -1,6 +1,5 @@
 import Link from "next/link";
 import React, { Component } from "react";
-import Image from "next/image";
 
 class Navbar extends Component {
     render() {
@@ -12,9 +11,9 @@ class Navbar extends Component {
                     <a>Home</a>
                 </Link>
 
-                {/*<a href="news">News</a>*/}
-                {/*<a href="contact">Contact</a>*/}
-                {/*<a href="about">About</a>*/}
+                <Link href="/users">
+                    <a>Users</a>
+                </Link>
 
                 <Link href="/login">
                     <a>Login</a>
@@ -22,10 +21,6 @@ class Navbar extends Component {
 
                 <Link href="/register">
                     <a>Register</a>
-                </Link>
-
-                <Link href="/users">
-                    <a>Users</a>
                 </Link>
 
                 <a href="javascript:void(0);" className="icon" onClick="myFunction()">
